Organized-PPT-master/requirements.txt
alabaster==0.7.12
attrs==19.3.0
Babel==2.8.0
blis==0.4.1
boto3==1.12.32
botocore==1.15.32
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
coverage==5.0.4
cycler==0.10.0
cymem==2.0.3
dask==2.13.0
decorator==4.4.2
dicttoxml==1.7.4
distributed==1.28.1
docutils==0.14
docx2txt==0.8
EbookLib==0.17.1
en-core-web-sm==2.2.0
entrypoints==0.3
extract-msg==0.23.1
Flask==1.1.1
future==0.18.2
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.12.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphviz==0.13.2
HeapDict==1.0.1
idna==2.9
imagesize==1.1.0
IMAPClient==2.1.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.0.2
kiwisolver==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
msgpack==1.0.0
murmurhash==1.0.2
networkx==2.4
nltk==3.4.3
np-utils==0.5.10.0
numexpr==2.7.0
numpy==1.16.4
numpydoc==0.9.1
olefile==0.46
packaging==20.3
pandas==0.24.2
pandocfilters==1.4.2
Pillow==5.2.0
pkg-resources==0.0.0
plac==0.9.6
preshed==3.0.2
protobuf==3.11.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycorenlp==0.3.0
pycparser==2.19
pycryptodome==3.9.0
pydot==1.4.1
pydotplus==2.0.2
pydub==0.23.1
pyemd==0.5.1
pyflakes==2.1.1
Pygments==2.6.1
PyMuPDF==1.16.13
pyparsing==2.4.6
PyPDF2==1.26.0
pyrsistent==0.16.0
pytextrank==2.0.1
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2019.3
PyYAML==5.3.1
regex==2017.4.5
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22
scipy==1.3.0
seaborn==0.10.0
six==1.14.0
sklearn==0.0
smart-open==1.10.0
snowballstemmer==1.2.1
sortedcontainers==2.1.0
spacy==2.2.4
SpeechRecognition==3.8.1
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
srsly==1.0.2
tblib==1.6.0
thinc==7.4.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.44.1
tzlocal==1.5.1
urllib3==1.25.8
wasabi==0.6.0
Werkzeug==1.0.0
wordcloud==1.5.0
XlsxWriter==1.2.8
yattag==1.13.2
zict==2.0.0
zipp==3.1.0
